scripts/requirements-hl.txt
# Hyperliquid trading agent dependencies
hyperliquid-python-sdk>=0.6.0
eth-account>=0.11.0
